Cleaned up Threads; still receiving out of memory error

+++ Concurrency Program/src/randomizers/RandomArray.java
@@ -4,24 +4,24 @@ import java.util.ArrayList;
 import java.util.Random;
 
 public class RandomArray {
-	private ArrayList<Long> randomArray;
+	private ArrayList<Integer> randomArray;
 	
 	public RandomArray (long arrayLength) {
 		setRandomArray(arrayLength);
 	}
 	
-	public ArrayList<Long> getRandomArray() {
+	public ArrayList<Integer> getRandomArray() {
 		return randomArray;
 	}
-	public void setRandomArray(ArrayList<Long> randomArray) {
+	public void setRandomArray(ArrayList<Integer> randomArray) {
 		this.randomArray = randomArray;
 	}
 	public void setRandomArray(long arrayLength) {
-		this.randomArray = new ArrayList<Long>();
+		this.randomArray = new ArrayList<Integer>();
 		Random randomGen = new Random();
 		
 		for(long counter = 0; counter < arrayLength; counter++) {
-			randomArray.add(randomGen.nextLong());
+			randomArray.add(randomGen.nextInt());
 		}
 	}
 }
